+++ tests/conftest.py
@@ -4,9 +4,3 @@ from django.conf import settings
 
 if not apps.ready and not settings.configured:
     setup()
-
-
-from testapp.models import TestModel
-
-
-TestModel.__test__ = False